+++ CHANGELOG.md
@@ -9,10 +9,13 @@ This is the log of notable changes to EAS CLI and related packages.
 ### 🎉 New features
 
 - Automatically create internal TestFlight group in EAS Submit command. ([#2839](https://github.com/expo/eas-cli/pull/2839) by [@evanbacon](https://github.com/evanbacon))
+- Sanitize and generate names for EAS Submit to prevent failures due to invalid characters or taken names. ([#2842](https://github.com/expo/eas-cli/pull/2842) by [@evanbacon](https://github.com/evanbacon))
 
 ### 🐛 Bug fixes
 
 ### 🧹 Chores
+
+- Make new autoIncremented builds start at nr 1 by default ([#2828](https://github.com/expo/eas-cli/pull/2828) by [@radoslawkrzemien](https://github.com/radoslawkrzemien))
 
 ## [14.5.0](https://github.com/expo/eas-cli/releases/tag/v14.5.0) - 2025-01-22